Description
The Meditation App is a simple project developed using JavaScript, CSS, and HTML. This project is an interesting project for developing our concentration power in meditation. The user can meditate and relax by enjoying the peaceful sound of a cloudy and sunny day. The user can adjust their meditation time from in the app.