Description
The Sketch Pad is a simple drawing application. This whole is in html5 and Javascript programming language. You can draw with the help of the cursor. You just can draw the stuff but unfortunately cannot save this in the system. This is a runtime program. But instead, you can save your work by downloading it as an image.